requirements.txt
Flask==3.1.0
python-dotenv==0.21.0
pypdfium2==4.30.1
PyPDF2==3.0.1
PyMuPDF==1.25.4
openpyxl==3.1.5
pandas==2.2.3
numpy
# Development dependencies
ipdb
pytest-playwright
pytest==8.3.5
pytest-asyncio